implemented a method to get enemy prefab to spawn according to prob weights.

+++ Assets/Scripts/EnemyWeights.cs
@@ -6,8 +6,8 @@ using UnityEngine;
 struct EnemyWeights {
     //Variable declaration
     //Note: I'm explicitly declaring them as public, but they are public by default. You can use private if you choose.
-    public int minProb;
-    public int maxProb;
+    public float minProb;
+    public float maxProb;
     public GameObject enemyPrefab;
     //Constructor (not necessary, but helpful)
     public EnemyWeights(int minProb, int maxProb, GameObject enemyPrefab) {
